requirements.txt
asn1crypto==1.4.0
azure-common==1.1.26
azure-core==1.9.0
azure-storage-blob==12.6.0
bleach==3.2.1
boto3==1.15.18
botocore==1.18.18
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
docutils==0.16
idna==2.10
importlib-metadata==3.1.1
isodate==0.6.0
jmespath==0.10.0
keyring==21.5.0
mock==4.0.3
msrest==0.6.19
numpy==1.19.4
oauthlib==3.1.0
oscrypto==1.2.1
packaging==20.7
pandas==1.1.4
pkginfo==1.6.1
providah==0.1.15.0
pyarrow==0.17.1
pycparser==2.20
pycryptodomex==3.9.9
Pygments==2.7.3
PyJWT==1.7.1
pyOpenSSL==20.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
readme-renderer==28.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
s3transfer==0.3.3
six==1.15.0
snowflake-connector-python==2.3.6
snowflake-sqlalchemy==1.2.4
SQLAlchemy==1.3.20
tqdm==4.54.1
twine==3.2.0
urllib3==1.25.11
webencodings==0.5.1
zipp==3.4.0
